KustoResultTable: parse results when you receive them. Keep 7th decimal digit of datetime/timespan(timedelta) (#96)
Make a long story short:
Python datetime/timedelta(timespan) native types supports 6 decimal digits.
Kusto supports 7 decimal digits, and pandas types supports 9.

Why am I telling this? We noticed that we don't give the existing accuracy to users that consumes pandas.
In order to do that we are keeping the 7th digit in the KustoResultRow class, and when converting into pandas we calculate back the 7th digit.
While doing that, we are now creating KustoResultRows on creation of the response, and not lazily.
The advantage in that is that one will receive True for the following:
r.primary_results[0][0] == r.primary_results[0][0]
In contrast to what we had so far.