Added connection pool capability
Database connection pool reuses connections which are already open in a thread-safe way and therefore avoids TCP and TLS handshake for each query.
requests.post has been replaced with urllib3.PoolManager.request which manages the connection pool to a single server.
Requirement for requests>=2.13.0 has been replaced with urllib3[secure]>=1.20. There was a significant improvement in urllib 1.20 so it is the earliest version which has to be demanded. The [secure] flavor adds dependency in certifi which is used for TLS certification validation.